requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.1
PyGithub==2.1.1
markdown==3.5.2
Werkzeug==3.0.1
Jinja2==3.1.3 